pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thenightops"
version = "0.2.0"
description = "Autonomous SRE Agent built on Google ADK and Remote MCP — with proactive detection, incident memory, and graduated remediation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "TheNightOps Contributors" },
]
keywords = ["sre", "agent", "mcp", "adk", "incident-response", "autonomous", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "google-adk>=0.3.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-container>=2.45.0",
    "kubernetes>=29.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "uvicorn>=0.29.0",
    "starlette>=0.37.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]
demo = [
    "fastapi>=0.110.0",
    "psutil>=5.9.0",
]

[project.scripts]
nightops = "thenightops.cli:app"

[tool.hatch.build.targets.wheel.force-include]
"src" = "thenightops"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
